Extract actual page/template components
  This way we can grab them without running graphql queries or importing
  styles
