+++ gatsby-config.js
@@ -55,7 +55,8 @@ module.exports = {
     {
       resolve: 'gatsby-plugin-netlify-cms',
       options: {
-        modulePath: `${__dirname}/src/cms/cms`
+        modulePath: `${__dirname}/src/cms/cms`,
+        stylesPath: `${__dirname}/src/components/layout.scss`
       }
     },
     {
