+++ src/components/events/index.tsx
@@ -1,0 +1,12 @@
+
+
+export interface IEventFields {
+  excerpt?: string,
+  html?: string,
+  frontmatter: {
+    contentType: 'event',
+    title: string,
+    date: string
+    link?: string
+  }
+}
