+++ src/components/events/summary.tsx
