+++ src/components/layout.scss
@@ -63,6 +63,8 @@
 
 @import "../colors";
 
+@import "footer/footer";
+
 // page-level layout
 
 html {
