+++ src/events/index.tsx
@@ -1,15 +1,4 @@
-
-
-export interface IEventFields {
-  excerpt?: string,
-  html?: string,
-  frontmatter: {
-    contentType: 'event',
-    title: string,
-    date: string
-    link?: string
-  }
-}
+import { graphql } from 'gatsby'
 
 export const eventFields = graphql`
 fragment eventFields on MarkdownRemark {
